multi_tool_agent/requirements.txt
google-generativeai>=0.3.0
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
python-dotenv>=1.0.0